requirements.txt
telethon
playwright
fastapi
uvicorn
python-dotenv
requests
